shift UI.java methods into GeneralUi.java and delete Output package containing UI.java
